refactor: put `presentation-screen.tsx` component along with its page
