+++ src/routes/present/partials/presentation-screen.tsx
